fix: posts 타입 변경, userId 명시,  게시글 사진 추가
