+++ json/readme.md
@@ -1,0 +1,10 @@
+PawsTimeProject1\package.json에
+
+"scripts": {
+"start:json": "json-server --watch json/db.json --port 4000"
+}
+
+추가,
+
+터미널에서
+pnpm start:json 로 json-server 실행
